# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AeroFS API",
    "description": "Remote file manager REST API. All endpoints require session authentication except public share endpoints.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/audit-logs": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Get system audit logs (Admin only)",
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/connections/{connection_id}/archive/compress": {
      "post": {
        "tags": [
          "archive"
        ],
        "summary": "Compress files into a ZIP or TAR.GZ archive",
        "operationId": "compress_files",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Archive created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/connections/{connection_id}/archive/entries": {
      "get": {
        "tags": [
          "archive"
        ],
        "summary": "List virtual directory contents inside an archive without full extraction",
        "operationId": "list_virtual_archive_endpoint",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archive_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subpath",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Virtual archive entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VirtualArchiveEntry"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/connections/{connection_id}/archive/extract": {
      "post": {
        "tags": [
          "archive"
        ],
        "summary": "Extract an archive into a target directory (Requires Create + Write permissions)",
        "operationId": "extract_archive_endpoint",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Archive extracted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/connections/{connection_id}/archive/extract-selected": {
      "post": {
        "tags": [
          "archive"
        ],
        "summary": "Extract specific selected entries from an archive into a destination directory (Requires Create + Write permissions)",
        "operationId": "extract_selected_archive_endpoint",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractSelectedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Selected entries extracted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/connections/{connection_id}/archive/read": {
      "get": {
        "tags": [
          "archive"
        ],
        "summary": "Stream or download a single entry directly from an archive",
        "operationId": "read_virtual_archive_entry_endpoint",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archive_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entry_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Virtual archive entry file content",
            "content": {
              "application/octet-stream": {}
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveOverwriteMode": {
        "type": "string",
        "enum": [
          "overwrite",
          "skip",
          "keep_both"
        ]
      },
      "ArchiveResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "entries_count": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "skipped_count": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "AuditLogEntry": {
        "type": "object",
        "required": [
          "id",
          "action",
          "status",
          "created_at"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "connection_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "ip_address": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "CompressRequest": {
        "type": "object",
        "required": [
          "base_path",
          "relative_paths",
          "destination_file"
        ],
        "properties": {
          "base_path": {
            "type": "string"
          },
          "destination_file": {
            "type": "string"
          },
          "format": {
            "type": [
              "string",
              "null"
            ]
          },
          "relative_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorCategory": {
        "type": "string",
        "enum": [
          "validation",
          "authentication",
          "authorization",
          "not_found",
          "conflict",
          "transfer_cancelled",
          "payload_too_large",
          "insufficient_storage",
          "rate_limited",
          "timeout",
          "provider",
          "io",
          "security",
          "internal"
        ]
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "category",
          "retryable",
          "message"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategory"
          },
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "user_action": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ExtractRequest": {
        "type": "object",
        "required": [
          "archive_path",
          "destination_dir"
        ],
        "properties": {
          "archive_path": {
            "type": "string"
          },
          "destination_dir": {
            "type": "string"
          },
          "format": {
            "type": [
              "string",
              "null"
            ]
          },
          "overwrite_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ArchiveOverwriteMode"
              }
            ]
          }
        }
      },
      "ExtractSelectedRequest": {
        "type": "object",
        "required": [
          "archive_path",
          "destination_dir",
          "entries"
        ],
        "properties": {
          "archive_path": {
            "type": "string"
          },
          "destination_dir": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "overwrite_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ArchiveOverwriteMode"
              }
            ]
          }
        }
      },
...
      "VirtualArchiveEntry": {
        "type": "object",
        "required": [
          "name",
          "path",
          "kind",
          "size"
        ],
        "properties": {
          "compressed_size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "kind": {
            "type": "string"
          },
          "modified_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aerofs_session"
      }
    }
  },
...
}